Fix timing bug in Mongodb replicaset configuration
Since start_server() and restart_server() do not actually verify whether the
port is open and receiving connections (even though mongo starts up,
pre-alloc/other setup takes a few minutes), the replicaset recipe was failing to
establish a connection with the db.  The change to the initd file waits for the
port to be open before proceeding, ensuring that the replicaset recipe will
work.

Fixes #17